Refactor out audio logic from spinner view
